Optimized db column size

+++ hedera-mirror-importer/src/main/resources/db/migration/v1/V1.58.2__ethereum_transaction.sql
@@ -24,7 +24,7 @@ create table if not exists ethereum_transaction
     signature_s              bytea    null,
     signature_v              bytea    null,
     to_address               bytea    null,
-    type                     int      not null,
+    type                     smallint not null,
     value                    bytea    null,
     primary key (consensus_timestamp)
 );
